cargo: extract full-profile key_system into smart-keymap-full-system-std
Move Nickel codegen of the Vec-backed full composite shell out of
smart-keymap so default builds and cargo doc no longer require nickel.
Cucumber, sizes, and the full-shell keymap tests depend on the new
workspace package instead.
