+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
     "keyberon-smart-keyboard",
     "rp2040-rtic-smart-keyboard",
+    "smart-keymap-full-system-std",
     "smart-keymap-macros",
     "smart-keymap-nickel-helper",
     "smart_keymap",
@@ -31,18 +32,9 @@ std = ["serde/std", "dep:serde_json"]
 name = "smart_keymap"
 crate-type = ["lib"]
 
-[[bin]]
-name = "sizes"
-required-features = ["std"]
-
 [[test]]
 name = "rust-integration"
 path = "tests/rust/keymap.rs"
-
-[[test]]
-name = "cucumber-keymap"
-path = "tests/cucumber/keymap.rs"
-harness = false
 
 [dependencies]
 heapless = { version = "0.8", features = ["serde"] }
@@ -54,9 +46,7 @@ serde = { version = "1.0", features = ["derive"], default-features = false }
 serde_json = { version = "1.0", optional = true }
 
 [dev-dependencies]
-cucumber = "0.21"
 ron = "0.8"
-futures = "0.3"
 
 smart-keymap-macros = { path = "./smart-keymap-macros" }
 smart-keymap-nickel-helper = { path = "./smart-keymap-nickel-helper" }
