+++ smart-keymap-full-system-std/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "smart-keymap-full-system-std"
+version.workspace = true
+license.workspace = true
+edition.workspace = true
+authors.workspace = true
+publish = false
+description = "Full-profile Vec-backed composite key_system for std consumers (cucumber, tooling)"
+
+[lib]
+name = "smart_keymap_full_system_std"
+path = "src/lib.rs"
+
+[[bin]]
+name = "sizes"
+path = "src/bin/sizes.rs"
+
+[[test]]
+name = "cucumber-keymap"
+path = "../tests/cucumber/keymap.rs"
+harness = false
+
+[dependencies]
+heapless = { version = "0.8", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+smart-keymap = { path = "..", features = ["std"] }
+
+[build-dependencies]
+smart-keymap-nickel-helper = { path = "../smart-keymap-nickel-helper" }
+
+[dev-dependencies]
+cucumber = "0.21"
+futures = "0.3"
+smart-keymap-nickel-helper = { path = "../smart-keymap-nickel-helper" }
